+++ src/SeablastController.php
@@ -483,10 +483,10 @@ class SeablastController
                 $this->identity->setTablePrefix($this->configuration->dbmsTablePrefix());
             }
             if (method_exists($this->identity, 'setCookiePath')) {
-                Debugger::barDump('Cookie Path injection');
+                Debugger::barDump('Cookie Path injection', 'Controller to Identity');
                 $this->identity->setCookiePath($this->getAppPath());
             } else {
-                Debugger::barDump('Cookie Path NOT INJECTED, i.e. remains default');
+                Debugger::barDump('Cookie Path NOT INJECTED, i.e. remains default', 'Controller to Identity');
             }
             Assert::methodExists($this->identity, 'isAuthenticated');
             if ($this->identity->isAuthenticated()) {