+++ src/Admin/AdminHelper.php
@@ -9,6 +9,7 @@ use Seablast\Seablast\SeablastConfiguration;
 use Seablast\Seablast\SeablastConstant;
 use Seablast\Seablast\Superglobals;
 use Tracy\Debugger;
+use Webmozart\Assert\Assert;
 
 /**
  * Helper methods for AdminModel, TableViewModel and ApiTableUpdateModel
@@ -103,6 +104,8 @@ class AdminHelper
 
         $columnTypes = [];
         while ($row = $result->fetch_object()) {
+            Assert::scalar($row->DATA_TYPE);
+            Assert::string($row->COLUMN_NAME);
             $columnTypes[$row->COLUMN_NAME] = (string) $row->DATA_TYPE;
         }
         $result->free();