+++ src/SeablastController.php
@@ -298,13 +298,16 @@ class SeablastController
         $clientIp = $server['REMOTE_ADDR'] ?? '';
 
         $proxyHeaders = (
-            (!empty($server['HTTP_X_FORWARDED_PROTO']) && strtolower($server['HTTP_X_FORWARDED_PROTO']) === 'https') ||
-            (!empty($server['HTTP_X_FORWARDED_SSL']) && strtolower($server['HTTP_X_FORWARDED_SSL']) === 'on')
-            );
+            (!empty($server['HTTP_X_FORWARDED_PROTO']) && is_string($server['HTTP_X_FORWARDED_PROTO'])
+              && strtolower($server['HTTP_X_FORWARDED_PROTO']) === 'https') ||
+            (!empty($server['HTTP_X_FORWARDED_SSL']) && is_string($server['HTTP_X_FORWARDED_SSL'])
+              && strtolower($server['HTTP_X_FORWARDED_SSL']) === 'on')
+        );
 
         return
-            (!empty($server['HTTPS']) && strtolower($server['HTTPS']) === 'on') ||
-            (!empty($server['REQUEST_SCHEME']) && strtolower($server['REQUEST_SCHEME']) === 'https') ||
+            (!empty($server['HTTPS']) && is_string($server['HTTPS']) && strtolower($server['HTTPS']) === 'on') ||
+            (!empty($server['REQUEST_SCHEME']) && is_string($server['REQUEST_SCHEME'])
+            && strtolower($server['REQUEST_SCHEME']) === 'https') ||
             (!empty($server['SERVER_PORT']) && $server['SERVER_PORT'] == '443') ||
             ($proxyHeaders && in_array($clientIp, $trustedProxies, true));
     }
@@ -473,16 +476,17 @@ class SeablastController
         // Authenticate: is there an identity manager to be used?
         if ($this->configuration->exists(SeablastConstant::SB_IDENTITY_MANAGER)) {
             $identityManager = $this->configuration->getString(SeablastConstant::SB_IDENTITY_MANAGER);
-            /* @phpstan-ignore-next-line Property $identity does not accept object. */
-            $this->identity = new $identityManager($this->configuration->mysqli());
+            /** @var IdentityManagerInterface $tempIdentity */
+            $tempIdentity = new $identityManager($this->configuration->mysqli());
+            $this->identity = $tempIdentity;
             if (method_exists($this->identity, 'setTablePrefix')) {
                 $this->identity->setTablePrefix($this->configuration->dbmsTablePrefix());
             }
             if (method_exists($this->identity, 'setCookiePath')) {
-                Debugger::barDump('Cookie Path injection');
+                Debugger::barDump('Cookie Path injection', 'Controller to Identity');
                 $this->identity->setCookiePath($this->getAppPath());
             } else {
-                Debugger::barDump('Cookie Path NOT INJECTED, i.e. remains default');
+                Debugger::barDump('Cookie Path NOT INJECTED, i.e. remains default', 'Controller to Identity');
             }
             Assert::methodExists($this->identity, 'isAuthenticated');
             if ($this->identity->isAuthenticated()) {