+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### `Fixed` for any bugfixes
 
 - fix: add assertions and typecastings to pass PHPStan static analysis at the `level: max` to make the code more robust
+- make sure that the database Tracy BarPanel is displayed when DbmsException is thrown in the model
 
 ### `Security` in case of vulnerabilities