+++ CHANGELOG.md
@@ -21,7 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.2.14] - 2026-01-04
 
-feat: database Tracy BarPanel is displayed for any Throwable
+feat: show database Tracy BarPanel for any Throwable
 
 ### Added